requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
email-validator>=2.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
python-dotenv==1.0.0

